[LabKey UX] Edit standard settings name/label/type - API Update
For the new Domain Designer, need to be able to retrieve and update a domain by the domain Id.  This has already been added to the standard Java and JS api.  Just need to add the parameters here to be passed through to the Java API.  Only making this available for the config object, not the parameter list.  Bumped version here.